// backend/sanity.types.ts
export type Product = {
  _id: string
  _type: 'product'
  _createdAt: string
  _updatedAt: string
  _rev: string
  sku?: string
  name?: string
  slug?: Slug
  description?: string
  manufacturer?: string
  category?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'category'
  }
  isESD?: boolean
  isEUItem?: boolean
  barcode?: string
  price?: {
    priceExclVAT?: number
    salePrice?: number
    specialPrice?: number
    specialSalePrice?: number
    startDate?: string
    endDate?: string
    vat?: number
    recycleTax?: number
    currency?: 'EUR' | 'RON' | 'MAD' | 'SAR' | 'AED'
  }
  volumePricing?: Array<{
    quantity?: number
    price?: number
    salePrice?: number
    discountPercent?: number
    _key: string
  }>
  variants?: Array<{
    id?: string
    title?: string
    price?: number
    listPrice?: number
    salePrice?: number
    inventory?: number
    ourInventory?: number
    barcodes?: string
    _key: string
  }>
  intelPoints?: number
  warranty?: string
  specifications?: Array<{
    name?: string
    value?: string
    _key: string
  }>
  inventory?: {
    quantity?: number
    ourInventory?: number
    purchaseOrders?: Array<{
      quantity?: number
      deliveryDate?: string
      _key: string
    }>
  }
  images?: Array<{
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
    _key: string
  }>
  lastUpdated?: string
}
// ...
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol